// src/lib/showfile/generated/DriverConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SacnDestination } from "./SacnDestination";

export type DriverConfig = { "kind": "null" } | { "kind": "ftdi", port: string | null, rateHz: number, } | { "kind": "enttecPro", port: string | null, universeIndex: number, rateHz: number, } | { "kind": "artNet", 
/**
 * Destination IP, e.g. "255.255.255.255" for broadcast or a node's
 * unicast address.
 */
destination: string, net: number, subnet: number, universe: number, rateHz: number, } | { "kind": "sacn", destination: SacnDestination, universe: number, priority: number, rateHz: number, };